+++ src/test/resources/edgeCases/insertGetIdModel.php
@@ -1,0 +1,10 @@
+<?php
+
+class User extends \Hyperf\Database\Model\Model {
+
+}
+
+User::query()->insertGetId([
+    'email' => 'john@example.com',
+    '<caret>' => 1,
+]);